arm: dma-removed: Add option to skip buffer zeroing
The DMA framework currently zeros all buffers because it (righfully so)
assumes that drivers will soon need to pass the memory to a device.
Some devices/use case may not require zeroed memory and there can
be an increase in performance if we skip the zeroing. Use the
DMA_ATTR_SKIP_ZEROING attr to allow skipping of DMA zeroing.

Change-Id: Ie7a8ae547834c29a6a4f327095e099c53526f3f7
Signed-off-by: Taniya Das <tdas@codeaurora.org>