defconfig: introduce cyanogenmod_kerneldebug_config
cyanogenmod_kerneldebug_defconfig is a defconfig fragment containing
debug related features which may not be appropriate to enable in
production builds due to their impact on performance or power.

This is not a full defconfig, just a list of deltas. It's meant
to be applied to a base defconfig using merge-config.sh.

Change-Id: Idbcddc8d7e5ab276272e1de7bfdeed88c0c6b2b9
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>